requirements.txt
pyyaml
anytree
vspec-tools
fuzzywuzzy
jinja2
